fix(stella-model): split OpenAI's Provider impl out of its god file, and regenerate the wire contract
Follow-ups inside the same change, both found by the gate rather than assumed:

- `check-file-size`: adding `Provider::model` grew `openai.rs` past its
  recorded ceiling, and a trait impl cannot be spread across modules — so the
  block that must absorb every future port method is the block that had to
  leave. `openai/provider.rs` now holds it: a pure delegation seam (`id`/`model`
  answer from fields, both completion methods forward to `complete_inner`), so
  nothing that reasons about SSE, pricing or the wire dialect moved. `openai.rs`
  ends 18 lines BELOW its ceiling rather than 4 above it, and no baseline entry
  was raised — a raised ceiling to turn a gate green is a defect, not a fix.
- `stella-cli/src/agent.rs` is likewise back to exactly its ceiling: the
  `stella init` model hint is one expression again, with the reasoning where it
  belongs (the trait's own doc) instead of duplicated at the call site. The
  five near-identical adapter doc comments went the same way.
- `wire-schema`: the new `UsageIncomplete::model` doc reaches
  `docs/wire/agentevent.d.ts` and both schemas. Regenerated and read as a wire
  change — it is additive only, a `description` on an existing `model: string`,
  with no field added, removed, renamed or re-typed.

`Provider::model` returning `None` is documented against its real instance:
`stella-serve`'s `RemoteProvider`, where the HOST picks the model and the
engine holds no ambient authority over it.

Refs #2831, #2952
