+++ crates/stella-model/src/anthropic.rs
@@ -791,8 +791,6 @@ impl Provider for AnthropicProvider {
         "anthropic"
     }
 
-    /// Bound to one model at construction, so a failed call can name it
-    /// (#2831).
     fn model(&self) -> Option<&str> {
         Some(&self.model)
     }
