+++ crates/stella-model/src/bedrock.rs
@@ -638,8 +638,6 @@ impl Provider for BedrockProvider {
         "bedrock"
     }
 
-    /// Bound to one model at construction, so a failed call can name it
-    /// (#2831).
     fn model(&self) -> Option<&str> {
         Some(&self.model)
     }
