+++ crates/stella-protocol/src/provider.rs
@@ -112,7 +112,10 @@ pub trait Provider: Send + Sync {
     /// swap were the rows unable to name what failed.
     ///
     /// `None` means the adapter is not bound to a single model and cannot
-    /// answer before dispatch, not that the model is unknowable in general;
+    /// answer before dispatch — `stella-serve`'s `RemoteProvider` is the real
+    /// instance, since there the HOST chooses the model and the engine holds
+    /// no ambient authority over it — not that the model is unknowable in
+    /// general;
     /// callers render it as [`UNKNOWN_MODEL`], which is the one documented
     /// spelling of "this envelope could not name its model". Defaulted rather
     /// than required so a test double or a future multiplexing adapter need
